modified read_engine and added write_engine
read_engine can be modified to rewrite the filename if it has a typo.
added option to exit from simulation
write_engine, for now, count how many char have each row.
the goal is to create a string as long as the max_row_length, read the file row by row and save only the numbers in the engine structure.
the "engine.P_max_sl = 100" has the only purpose to check if the engine passed by the function, has the modified value. It will be eliminated in the future.